PlayerTab was adjusted to automatically set its tab size based on the width of its tab name. Tab was adjusted to only render its name at integers to avoid blurriness. WorldExporer now leaves the prefixes, materials, players, and itemBase list frames collapsed. TabState no longer auto-makes tabs and instead just shows the WorldExplorer.

+++ GM Calc 2/src/org/gmcalc2/gui/Tab.java
@@ -106,7 +106,7 @@ public class Tab extends GUISubcontext {
 		g.fill(tabShape);
 		g.setColor(tabNameColor);
 		g.setFont(font);
-		g.drawString(tabName, tabShape.getX() + tabShape.getWidth() / 2 - (font.getWidth(tabName) / 2), y1);
+		g.drawString(tabName, (int)(tabShape.getX() + tabShape.getWidth() / 2 - (font.getWidth(tabName) / 2)), y1);
 		
 		// Render the interior if enabled.
 		if (enabled)
